Can install espresso on OS/X
Espresso uses unix utility `find dir -name \*.x` to determine the executable to
install. On OS/X, it also finds a bunch of debug symbols associated
with the exectuble. These files have the same name, but are in a
different directory. So the install process is done by hand in the
package.py file, for OS/X.