Clang & Boost: No PCH (#2827)
Although it hurts a little, officially pre-compiled headers in
boost are only supported for gcc & msvc and the latest clang
releases still fail to build boost with it.

Therefore, I disabled building those to get boost build with
clang 3.9.0 on an Ubuntu 14.04 (x86).

Links to documentation and boost bug reports are inline, so
people can later on check if they still apply. Seems just to
be a bug in `Boost.Build` that tries to set `-o` with multiple
output files.