Update package: add info for emacs@26.1 (#8302)
* Update package: add info for emacs@26.1

In addition to the digest, it needs gnutls (perhaps there's a way to
turn it off, but it seems like a Good Thing to encourage, so...).

* Clean up tls variant handling

- Don't to specify the dependency for @26.1:, just need to fix what's
  already there.
- Don't assume that configure will DTRT w.r.t. gnutls when it's
  discovered in the environmnet, use explicit configure switches.
- Clean up some unnecessary dependencies when +tls.  Why would you
  need libxpm when building +tls?  Will ask the original author about
  this.

* Use append when adding configure switches

* Add dependency on pcre

When one builds +X, Spack's pcre is pulled into the build via
the X dependencies.  When one builds without it, the build discovers
and links against the system library.

We might as well just depend on it and have a consistent outcome.